requirements.txt
bitarray==2.8.3
mmh3==4.0.1
pycryptodome==3.19.0
pympler==1.1